Create a small TournName component to render the tournament name in the Bracket. Also fix invite reducer to correctly toggle.
