+++ client/src/components/RosterContainer.js
@@ -5,8 +5,6 @@ import Roster from './roster/Roster.jsx';
 
 
 function mapStateToProps(state){
-  console.log('mapStateToProps showTournList', state.getIn(['header','showTournList']));
-  console.log('mapStateToProps userTourns', state.getIn(['header', 'userData', 'userTourns']));
   console.log(state.get('tournament').get('roster').toJS());
   console.log(state.getIn(['tournament', 'roster']).toJS());
   return {
