+++ client/src/reducers/info.jsx
@@ -11,6 +11,8 @@ function handleUpdateType(state, tournType) {
 }
 
 function handleUpdateName(state, tournName) {
+  console.log('Info reducer: handleUpdateName');
+  console.log('handleUpdateName: tournName =', tournName);
   return state.set('tournName', tournName);
 }
 
@@ -21,13 +23,13 @@ function handleUpdateRules(state, rules) {
 export default function info(state = {}, action) {
   switch (action.type) {
     case 'UPDATE_ID':
-      return handleUpdateId(state.tournInfo, action.tournId);
+      return handleUpdateId(state, action.tournId);
     case 'UPDATE_TYPE':
-      return handleUpdateType(state.tournInfo, action.tournType);
+      return handleUpdateType(state, action.tournType);
     case 'UPDATE_NAME':
-      return handleUpdateName(state.tournInfo, action.tournName);
+      return handleUpdateName(state, action.tournName);
     case 'UPDATE_RULES':
-      return handleUpdateRules(state.tournInfo, action.rules);
+      return handleUpdateRules(state, action.rules);
     default:
       return state;
   }
